+++ apps/www/lib/telemetry.ts
@@ -1,7 +1,7 @@
-import { post } from '~/lib/fetchWrapper'
-import { API_URL, IS_PROD, IS_PREVIEW } from 'lib/constants'
+import { LOCAL_STORAGE_KEYS, TelemetryProps } from 'common'
+import { API_URL, IS_PREVIEW, IS_PROD } from 'lib/constants'
 import { NextRouter } from 'next/router'
-import { isBrowser, LOCAL_STORAGE_KEYS } from 'common'
+import { post } from '~/lib/fetchWrapper'
 
 export interface TelemetryEvent {
   category: string
@@ -10,17 +10,12 @@ export interface TelemetryEvent {
   value?: string
 }
 
-export interface TelemetryProps {
-  screenResolution?: string
-  language: string
-}
-
 const noop = () => {}
 
 // This event is the same as in studio/lib/telemetry.tx
 // but uses different ENV variables for www
 
-const sendEvent = (event: TelemetryEvent, gaProps: TelemetryProps, router: NextRouter) => {
+const sendEvent = (event: TelemetryEvent, telemetryProps: TelemetryProps, router: NextRouter) => {
   const consent =
     typeof window !== 'undefined'
       ? localStorage.getItem(LOCAL_STORAGE_KEYS.TELEMETRY_CONSENT)
@@ -35,20 +30,22 @@ const sendEvent = (event: TelemetryEvent, gaProps: TelemetryProps, router: NextR
   const title = typeof document !== 'undefined' ? document?.title : ''
   const referrer = typeof document !== 'undefined' ? document?.referrer : ''
 
+  const { page_url, search, language, viewport_height, viewport_width } = telemetryProps
+
   return post(
     `${API_URL}/telemetry/event`,
     {
+      page_url,
       action: action,
-      page_url: window.location.href,
       page_title: title,
       pathname: router.pathname,
       ph: {
+        search,
         referrer,
-        language: router?.locale ?? 'en-US',
+        language,
+        viewport_height,
+        viewport_width,
         user_agent: navigator.userAgent,
-        search: window.location.search,
-        viewport_height: isBrowser ? window.innerHeight : 0,
-        viewport_width: isBrowser ? window.innerWidth : 0,
       },
       custom_properties: { category, label, value } as any,
     },
